requirements.txt
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
boto3==1.36.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels-redis==4.2.1
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
dj-database-url==3.1.2
Django==5.2.8
django-allauth==65.15.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-unfold==0.45.0
django-ratelimit==4.1.0
django-storages[s3]==1.14.4
gunicorn==25.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dotenv==1.2.1
razorpay==2.0.0
redis==5.2.1
requests==2.32.5
service-identity==24.2.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
whitenoise==6.12.0
zope.interface==8.2
